Split Asserter into partial classes
